[Bronze V] Title: 이상한 기호, Time: 84 ms, Memory: 11492 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -7,13 +7,13 @@ public class Main {
 
         StringTokenizer st = new StringTokenizer(br.readLine());
 
-        int a = Integer.parseInt(st.nextToken());
-        int b = Integer.parseInt(st.nextToken());
+        long a = Integer.parseInt(st.nextToken());
+        long b = Integer.parseInt(st.nextToken());
 
         System.out.println(cal(a, b));
     }
 
-    public static int cal(int a, int b) {
+    public static long cal(long a, long b) {
         return (a + b) * (a - b);
     }
 }
